operations are separate from commands now
