+++ webdriver/src/commands/find_elements.ts
@@ -46,7 +46,7 @@ async function findElementsRequest(
 		{
 			method: "POST",
 			headers,
-			body: JSON.stringify({ using: "css selector", value: css_selector}),
+			body: JSON.stringify({ using: "css selector", value: css_selector }),
 			signal,
 		},
 	);
