+++ webdriver/src/commands/mod.ts
@@ -1,5 +1,5 @@
-// LOG not technically a webdriver command but needed for degbug
-export * from "./add_cookie.js";
+// An implementation of a subset of webdriver spec actions.
+// LOG is not technically a webdriver command but needed for some degbug.
 export * from "./element_click.js";
 export * from "./element_send_keys.js";
 export * from "./find_element_from_element.js";
@@ -10,8 +10,6 @@ export * from "./find_elements_from_shadow_root.js";
 export * from "./find_elements.js";
 export * from "./get_element_shadow_root.js";
 export * from "./log.js";
-export * from "./navigate_to.js";
-export * from "./sessions.js";
 export * from "./take_element_screenshot.js";
 
 // should errors be sent to session errors
