+++ webdriver/src/operations/add_cookie.ts
@@ -1,6 +1,6 @@
 import type { WebdriverParams } from "../config.js";
 
-import { headers } from "./flyweight.js";
+import { headers } from "../commands/flyweight.js";
 
 export async function addCookie(
 	params: WebdriverParams,
