+++ webdriver/src/operations/serve_file.ts
@@ -1,12 +1,10 @@
-import type { WebdriverParams } from "./config.js";
 import type { IncomingMessage, ServerResponse } from "http";
 
 import * as fs from "fs";
 import * as path from "path";
 
-let headers = new Headers([["Content-Type", "application/json"]]);
 let cwd = process.cwd();
-const parentPath = path.join(import.meta.url.substring(5), "../../../");
+const parentPath = path.join(import.meta.url.substring(5), "../../../../");
 
 const MIME_TYPES: Record<string, string> = {
 	octet: "application/octet-stream",
@@ -19,43 +17,6 @@ const MIME_TYPES: Record<string, string> = {
 	ico: "image/x-icon",
 	svg: "image/svg+xml",
 };
-
-export async function untilWebdriverReady(
-	params: WebdriverParams,
-	signal: AbortSignal | undefined,
-): Promise<void> {
-	let { url } = params;
-
-	while (signal && !signal.aborted) {
-		try {
-			// this should be a command
-			let res = await fetch(new URL("/status", url), {
-				method: "GET",
-				headers,
-				body: null,
-				signal,
-			});
-
-			if (200 === res.status) {
-				let json = await res.json();
-				let { ready } = json?.value;
-				if (typeof ready === "boolean" && ready) return;
-			}
-		} catch {}
-
-		await sleep(30);
-	}
-
-	throw new Error("Webdriver was never ready.");
-}
-
-function sleep(timeMs: number): Promise<void> {
-	return new Promise(function (resolve) {
-		setTimeout(function () {
-			resolve();
-		}, timeMs);
-	});
-}
 
 export async function serveFile(req: IncomingMessage, res: ServerResponse) {
 	let { url, method } = req;
