+++ webdriver/src/operations/until_ready.ts
@@ -1,0 +1,39 @@
+import type { WebdriverParams } from "../config.js";
+
+let headers = new Headers([["Content-Type", "application/json"]]);
+
+export async function untilWebdriverReady(
+	params: WebdriverParams,
+	signal: AbortSignal | undefined,
+): Promise<void> {
+	let { url } = params;
+
+	while (signal && !signal.aborted) {
+		try {
+			let res = await fetch(new URL("/status", url), {
+				method: "GET",
+				headers,
+				body: null,
+				signal,
+			});
+
+			if (200 === res.status) {
+				let json = await res.json();
+				let { ready } = json?.value;
+				if (typeof ready === "boolean" && ready) return;
+			}
+		} catch {}
+
+		await sleep(30);
+	}
+
+	throw new Error("Webdriver was never ready.");
+}
+
+function sleep(timeMs: number): Promise<void> {
+	return new Promise(function (resolve) {
+		setTimeout(function () {
+			resolve();
+		}, timeMs);
+	});
+}
